+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- fix editing for link fields without a title value
+
 ### Security
 
 ## [0.20.0] - 2025-07-08